now this code is making it render 5-6 times
The code is being rendered multiple times because the useFetch hook is being called on every render. This is because the useFetch hook is called in the body of the App component, which gets called on every render.

To make it render only once, you can move the useFetch hook call into a useEffect hook with an empty dependency array. This way, the hook will only be called once, when the component is mounted.